# Conflicts:
#	nodedoc/analyticsdoc.js
#	nodedoc/analyticsdoc.js.html
#	nodedoc/fsdoc.js
#	nodedoc/ladoc.js
#	nodedoc/ladoc_structures.js
#	nodedoc/module-analytics-metrics.ClassificationScore.html
#	nodedoc/module-analytics-metrics.MeanAbsoluteError.html
#	nodedoc/module-analytics-metrics.MeanAbsolutePercentageError.html
#	nodedoc/module-analytics-metrics.MeanError.html
#	nodedoc/module-analytics-metrics.MeanSquareError.html
#	nodedoc/module-analytics-metrics.PredictionCurve.html
#	nodedoc/module-analytics-metrics.R2Score.html
#	nodedoc/module-analytics-metrics.RootMeanSquareError.html
#	nodedoc/module-analytics-metrics.html
#	nodedoc/module-analytics-preprocessing.Binarizer.html
#	nodedoc/module-analytics-preprocessing.html
#	nodedoc/module-analytics.KMeans.html
#	nodedoc/module-analytics.OneVsAll.html
#	nodedoc/module-analytics.PCA.html
#	nodedoc/module-analytics.html
#	nodedoc/module-la.html
#	nodedoc/module-qm.CircularRecordBuffer.html
#	nodedoc/module-qm.FeatureSpace.html
#	nodedoc/module-qm.Iterator.html
#	nodedoc/module-qm.Record.html
#	nodedoc/module-qm.RecordSet.html
#	nodedoc/module-qm.html
#	nodedoc/qminerdoc.js
#	nodedoc/qminerdoc.js.html
#	nodedoc/scripts/jsdoc-toc.js
#	src/nodejs/analytics/analytics.cpp
#	src/nodejs/intellisense/analytics_intellisense.js
#	src/nodejs/intellisense/fs_intellisense.js
#	src/nodejs/intellisense/la_intellisense.js
#	src/nodejs/intellisense/qminer_intellisense.js
#	test/nodejs/exampleanalyticsdoc.js

+++ src/glib/base/json.h
@@ -114,7 +114,7 @@ public:
   bool GetBool() const {EAssert(IsBool()); return Bool;}
   double GetNum() const {EAssert(IsNum()); return Num;}
   int GetInt() const {EAssert(IsNum()); return TFlt::Round(Num);}
-  uint64 GetUInt64() const {EAssert(IsNum()); return uint64(Num);}
+  uint64 GetUInt64() const {EAssert(IsNum()); return (unsigned)(int64)(Num);}
   int64 GetInt64() const { EAssert(IsNum()); return int64(Num); }
   const TStr& GetStr() const {EAssert(IsStr()); return Str;}
   uint64 GetTmMSecs() const { return TTm::GetMSecsFromTm(GetTm()); }
